V1.12 : Aggiunto db locale da qualche parte e recupero password funzionante

+++ app/src/main/java/com/example/seasonapp/Recensioni.kt
@@ -1,5 +1,5 @@
 package com.example.seasonapp
 
-class Recensioni(val testo: String, val stelle : Int){
+class Recensioni(val testo: String, val stelle: Float){
     //card
 }
